monsters move and trigger attacks properly
human attacks first if moves into monster,
monster attacks first if moves into player (happens rarely, usually only
when many monsters surrounding player, as the monster has to move once
to keep up with the player if he the player moves to a new '.' spot)

+++ town.cpp
@@ -402,7 +402,7 @@ void Town::handleInventory() {
       case 6:
       case 7:
       case 8:
-      case 9: if (inputInt < currItemList.size()) {
+      case 9: if (inputInt < (int)currItemList.size()) {
                 currItemList.at(inputInt).equip(thePlayer);
                 thePlayer->setItems(currItemList);
               }
